added sqlite controller

+++ backend/domain/src/commonMain/kotlin/Controller.kt
@@ -5,9 +5,9 @@ interface Controller {
 
     fun addTask(task: Task)
 
-    fun removeTask(id: Int): Boolean
+    fun removeTask(id: Long): Boolean
 
     fun tasksByPriority(priority: Priority): List<IdTask>
 
-    fun taskById(taskId: Int): IdTask?
+    fun taskById(taskId: Long): IdTask?
 }
