+++ backend/presenter/public/src/commonMain/kotlin/DTO.kt
@@ -14,7 +14,7 @@ data class DTOTask(
 
 @Serializable
 data class DTOIdTask(
-    override val id: Int,
+    override val id: Long,
     override val name: String,
     override val description: String,
     override val priority: Priority,
