+++ backend/presenter/public/src/commonMain/kotlin/Presenter.kt
@@ -3,11 +3,11 @@ package hernanbosqued.backend.presenter
 interface Presenter {
     fun allTasks(): List<DTOIdTask>
 
-    fun taskById(taskId: Int?): Result<DTOIdTask, StatusCode>
+    fun taskById(taskId: Long?): Result<DTOIdTask, StatusCode>
 
     fun taskByPriority(priorityStr: String?): Result<List<DTOIdTask>, StatusCode>
 
     fun addTask(task: DTOTask): Result<Unit, StatusCode>
 
-    fun removeTask(taskId: Int?): Result<Unit, StatusCode>
+    fun removeTask(taskId: Long?): Result<Unit, StatusCode>
 }
